# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Criteria {
  id: number;
  targetRoles: string[];
  industries: string[];
  /** @nullable */
  minSalary?: number | null;
  locations: string[];
  mustHave: string[];
  niceToHave: string[];
  avoid: string[];
  yourName: string;
  yourEmail: string;
  updatedAt: string;
}

export interface UpdateCriteriaBody {
  targetRoles?: string[];
  industries?: string[];
  /** @nullable */
  minSalary?: number | null;
  locations?: string[];
  mustHave?: string[];
  niceToHave?: string[];
  avoid?: string[];
  yourName?: string;
  yourEmail?: string;
}

export interface Company {
  id: number;
  name: string;
  atsType: string;
  /** @nullable */
  atsSlug?: string | null;
  /** @nullable */
  careersUrl?: string | null;
  createdAt: string;
}

export type CreateCompanyBodyAtsType =
  (typeof CreateCompanyBodyAtsType)[keyof typeof CreateCompanyBodyAtsType];

export const CreateCompanyBodyAtsType = {
  greenhouse: "greenhouse",
  lever: "lever",
  workday: "workday",
  other: "other",
} as const;

export interface CreateCompanyBody {
  name: string;
  atsType: CreateCompanyBodyAtsType;
  /** @nullable */
  atsSlug?: string | null;
  /** @nullable */
  careersUrl?: string | null;
}

export interface Job {
  id: number;
  title: string;
  company: string;
  location: string;
  /** @nullable */
  salary?: string | null;
  applyUrl: string;
  whyGoodFit: string;
  matchScore: number;
  status: string;
  /** @nullable */
  tailoredResume?: string | null;
  /** @nullable */
  coverLetter?: string | null;
  /** @nullable */
  scoutRunId?: number | null;
  foundAt: string;
}

export interface UpdateJobStatusBody {
  status: string;
}

export interface Resume {
  id: number;
  content: string;
  updatedAt: string;
}

export interface UpdateResumeBody {
  content: string;
}

export interface ScoutRun {
  id: number;
  status: string;
  jobsFound: number;
  emailSent: boolean;
  /** @nullable */
  error?: string | null;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface ScoutRunResult {
  runId: number;
  jobsFound: number;
  emailSent: boolean;
  message: string;
}

export interface GmailStatus {
  connected: boolean;
  /** @nullable */
  email?: string | null;
  message: string;
}

export interface GmailSetupUrl {
  url: string;
  instructions: string;
}

export interface SendDigestResult {
  success: boolean;
  message: string;
  jobsSent: number;
}

export type ListJobsParams = {
  status?: string;
};

export type GmailCallbackParams = {
  code?: string;
  error?: string;
};