# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Get job search criteria
 */
export const GetCriteriaResponse = zod.object({
  id: zod.number(),
  targetRoles: zod.array(zod.string()),
  industries: zod.array(zod.string()),
  minSalary: zod.number().nullish(),
  locations: zod.array(zod.string()),
  mustHave: zod.array(zod.string()),
  niceToHave: zod.array(zod.string()),
  avoid: zod.array(zod.string()),
  yourName: zod.string(),
  yourEmail: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary Update job search criteria
 */
export const UpdateCriteriaBody = zod.object({
  targetRoles: zod.array(zod.string()).optional(),
  industries: zod.array(zod.string()).optional(),
  minSalary: zod.number().nullish(),
  locations: zod.array(zod.string()).optional(),
  mustHave: zod.array(zod.string()).optional(),
  niceToHave: zod.array(zod.string()).optional(),
  avoid: zod.array(zod.string()).optional(),
  yourName: zod.string().optional(),
  yourEmail: zod.string().optional(),
});

export const UpdateCriteriaResponse = zod.object({
  id: zod.number(),
  targetRoles: zod.array(zod.string()),
  industries: zod.array(zod.string()),
  minSalary: zod.number().nullish(),
  locations: zod.array(zod.string()),
  mustHave: zod.array(zod.string()),
  niceToHave: zod.array(zod.string()),
  avoid: zod.array(zod.string()),
  yourName: zod.string(),
  yourEmail: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary List target companies
 */
export const ListCompaniesResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  atsType: zod.string(),
  atsSlug: zod.string().nullish(),
  careersUrl: zod.string().nullish(),
  createdAt: zod.date(),
});
export const ListCompaniesResponse = zod.array(ListCompaniesResponseItem);

/**
 * @summary Add a target company
 */
export const CreateCompanyBody = zod.object({
  name: zod.string(),
  atsType: zod.enum(["greenhouse", "lever", "workday", "other"]),
  atsSlug: zod.string().nullish(),
  careersUrl: zod.string().nullish(),
});

/**
 * @summary Remove a target company
 */
export const DeleteCompanyParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary List job matches
 */
export const ListJobsQueryParams = zod.object({
  status: zod.coerce.string().optional(),
});

export const ListJobsResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  location: zod.string(),
  salary: zod.string().nullish(),
  applyUrl: zod.string(),
  whyGoodFit: zod.string(),
  matchScore: zod.number(),
  status: zod.string(),
  tailoredResume: zod.string().nullish(),
  coverLetter: zod.string().nullish(),
  scoutRunId: zod.number().nullish(),
  foundAt: zod.date(),
});
export const ListJobsResponse = zod.array(ListJobsResponseItem);

/**
 * @summary Get a job match
 */
export const GetJobParams = zod.object({
  id: zod.coerce.number(),
});

export const GetJobResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  location: zod.string(),
  salary: zod.string().nullish(),
  applyUrl: zod.string(),
  whyGoodFit: zod.string(),
  matchScore: zod.number(),
  status: zod.string(),
  tailoredResume: zod.string().nullish(),
  coverLetter: zod.string().nullish(),
  scoutRunId: zod.number().nullish(),
  foundAt: zod.date(),
});

/**
 * @summary Update job status
 */
export const UpdateJobStatusParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateJobStatusBody = zod.object({
  status: zod.string(),
});

export const UpdateJobStatusResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  location: zod.string(),
  salary: zod.string().nullish(),
  applyUrl: zod.string(),
  whyGoodFit: zod.string(),
  matchScore: zod.number(),
  status: zod.string(),
  tailoredResume: zod.string().nullish(),
  coverLetter: zod.string().nullish(),
  scoutRunId: zod.number().nullish(),
  foundAt: zod.date(),
});

/**
 * @summary Generate tailored resume and cover letter for a job
 */
export const GenerateJobDocsParams = zod.object({
  id: zod.coerce.number(),
});

export const GenerateJobDocsResponse = zod.object({
  id: zod.number(),
  title: zod.string(),
  company: zod.string(),
  location: zod.string(),
  salary: zod.string().nullish(),
  applyUrl: zod.string(),
  whyGoodFit: zod.string(),
  matchScore: zod.number(),
  status: zod.string(),
  tailoredResume: zod.string().nullish(),
  coverLetter: zod.string().nullish(),
  scoutRunId: zod.number().nullish(),
  foundAt: zod.date(),
});

/**
 * @summary Get the stored resume
 */
export const GetResumeResponse = zod.object({
  id: zod.number(),
  content: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary Update the resume
 */
export const UpdateResumeBody = zod.object({
  content: zod.string(),
});

export const UpdateResumeResponse = zod.object({
  id: zod.number(),
  content: zod.string(),
  updatedAt: zod.date(),
});

/**
 * @summary Run the job scout agent manually
 */
export const RunScoutResponse = zod.object({
  runId: zod.number(),
  jobsFound: zod.number(),
  emailSent: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get scout run history
 */
export const GetScoutStatusResponseItem = zod.object({
  id: zod.number(),
  status: zod.string(),
  jobsFound: zod.number(),
  emailSent: zod.boolean(),
  error: zod.string().nullish(),
  startedAt: zod.date(),
  completedAt: zod.string().nullish(),
});
export const GetScoutStatusResponse = zod.array(GetScoutStatusResponseItem);

/**
 * @summary Check Gmail connection status
 */
export const GetGmailStatusResponse = zod.object({
  connected: zod.boolean(),
  email: zod.string().nullish(),
  message: zod.string(),
});

/**
 * @summary Get Gmail OAuth setup URL
 */
export const GetGmailSetupUrlResponse = zod.object({
  url: zod.string(),
  instructions: zod.string(),
});

/**
 * @summary Handle Gmail OAuth callback
 */
export const GmailCallbackQueryParams = zod.object({
  code: zod.coerce.string().optional(),
  error: zod.coerce.string().optional(),
});

export const GmailCallbackResponse = zod.object({
  connected: zod.boolean(),
  email: zod.string().nullish(),
  message: zod.string(),
});

/**
 * @summary Disconnect Gmail
 */
export const DisconnectGmailResponse = zod.object({
  connected: zod.boolean(),
  email: zod.string().nullish(),
  message: zod.string(),
});

/**
 * @summary Send job digest email manually
 */
export const SendDigestEmailResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  jobsSent: zod.number(),
});