# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  atsType: string;
  /** @nullable */
  atsSlug?: string | null;
  /** @nullable */
  careersUrl?: string | null;
  createdAt: Date;
}