# lib/api-zod/src/generated/types/createCompanyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateCompanyBodyAtsType } from "./createCompanyBodyAtsType";

export interface CreateCompanyBody {
  name: string;
  atsType: CreateCompanyBodyAtsType;
  /** @nullable */
  atsSlug?: string | null;
  /** @nullable */
  careersUrl?: string | null;
}