# lib/api-zod/src/generated/types/gmailSetupUrl.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface GmailSetupUrl {
  url: string;
  instructions: string;
}