# lib/api-zod/src/generated/types/scoutRun.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoutRun {
  id: number;
  status: string;
  jobsFound: number;
  emailSent: boolean;
  /** @nullable */
  error?: string | null;
  startedAt: Date;
  /** @nullable */
  completedAt?: string | null;
}