# lib/api-zod/src/generated/types/updateCriteriaBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Job Scout Agent API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateCriteriaBody {
  targetRoles?: string[];
  industries?: string[];
  /** @nullable */
  minSalary?: number | null;
  locations?: string[];
  mustHave?: string[];
  niceToHave?: string[];
  avoid?: string[];
  yourName?: string;
  yourEmail?: string;
}